Added memory functions, search and delete

+++ hashtable.h
@@ -6,7 +6,7 @@ class HashTable {
 public:
     HashTable(int size);
     bool isEmpty();
-    void insert(string key, int value);
+    void insert(string key, double value);
     int search(string key);
     void remove(string key); // delete is a keyword
     void printTable();
